refactor: separate parser from dot lint policy
